added code for printing text to screen
Some place holder example code in place. Needs to be replaced with the sstream solution. Should actually consider using a map instead of a vector to hold the font_rectangle as currently I will need to insert values by number. A map would require a font_manager class which is a bit of a hassle. For now I will just have to use comments to keep tract of which font_rectangle should have which text inserted into it.